Moved pareto series to subfolder
